docs(Slug): move slug examples under experimental section
