+++ packages/react/src/components/TextArea/TextArea-slug.stories.js
@@ -1,0 +1,112 @@
+import React from 'react';
+import Button from '../Button';
+import TextArea from './';
+import { IconButton } from '../IconButton';
+import { View, FolderOpen, Folders } from '@carbon/icons-react';
+import { Slug, SlugContent, SlugActions } from '../Slug';
+
+export default {
+  title: 'Experimental/unstable__Slug/TextArea',
+  component: null,
+};
+
+const textareaProps = {
+  labelText: 'Text Area label',
+  className: 'some-class',
+  placeholder: 'Placeholder text',
+  id: 'test5',
+  rows: 4,
+};
+
+export const _WithSlug = (args) => {
+  const { revertActive, ...rest } = args;
+  const slug = (
+    <Slug className="slug-container" revertActive={revertActive}>
+      <SlugContent>
+        <div>
+          <p className="secondary">AI Explained</p>
+          <h1>84%</h1>
+          <p className="secondary bold">Confidence score</p>
+          <p className="secondary">
+            Lorem ipsum dolor sit amet, di os consectetur adipiscing elit, sed
+            do eiusmod tempor incididunt ut fsil labore et dolore magna aliqua.
+          </p>
+          <hr />
+          <p className="secondary">Model type</p>
+          <p className="bold">Foundation model</p>
+        </div>
+        <SlugActions>
+          <IconButton kind="ghost" label="View">
+            <View />
+          </IconButton>
+          <IconButton kind="ghost" label="Open Folder">
+            <FolderOpen />
+          </IconButton>
+          <IconButton kind="ghost" label="Folders">
+            <Folders />
+          </IconButton>
+          <Button>View literature</Button>
+        </SlugActions>
+      </SlugContent>
+    </Slug>
+  );
+
+  return <TextArea {...textareaProps} slug={slug} {...rest} />;
+};
+
+_WithSlug.args = {
+  revertActive: false,
+  invalid: false,
+  invalidText:
+    'Error message that is really long can wrap to more lines but should not be excessively long.',
+  disabled: false,
+  warn: false,
+  warnText:
+    'Warning message that is really long can wrap to more lines but should not be excessively long.',
+};
+
+_WithSlug.argTypes = {
+  children: {
+    table: {
+      disable: true,
+    },
+  },
+  className: {
+    table: {
+      disable: true,
+    },
+  },
+  disabled: {
+    control: {
+      type: 'boolean',
+    },
+  },
+  invalid: {
+    control: {
+      type: 'boolean',
+    },
+  },
+  invalidText: {
+    control: {
+      type: 'text',
+    },
+  },
+  warn: {
+    control: {
+      type: 'boolean',
+    },
+  },
+  warnText: {
+    control: {
+      type: 'text',
+    },
+  },
+  revertActive: {
+    control: {
+      type: 'boolean',
+    },
+    table: {
+      category: 'Slug',
+    },
+  },
+};
